LockingFilesystemStore: Add SharedFileLockock and fix transactions.
Added SharedFileLock, which is a shared (non-exclusive/read) lock.
Use SharedFileLock for reads in LockingFileSystemStore.
Acquire exclusive locks on all access within a LockingFilesystemStore
transaction. This prevents stale writes to the keys in event of
simultaneous writes, but also may potentially have deadlocks in case
the sequence of writes between clients are different.