ENH: thread pool executor: add `wait_at_exit` option.
Sometimes, it is useful to execute non-essential tasks in
the `EnhancedThreadPoolExecutor` which can be safely killed and
should not prevent the application from exiting, for example
fetching updates information or background indexing of files.
This commit add an optional argument `wait_at_exit` which can
be set to `False` to kill the worker threads abruptly at
interpreter shutdown instead of waiting for them to finish.